Switch to correct BorrowedDto
