+++ app/src/main/java/com/book/library/book/BookRestController.java
@@ -3,6 +3,7 @@ package com.book.library.book;
 import com.book.library.dto.AnyBookDto;
 import com.book.library.dto.AvailableBookDto;
 import com.book.library.dto.BorrowedBookDto;
+import com.book.library.dto.BorrowedDto;
 import io.micrometer.observation.Observation;
 import io.micrometer.observation.ObservationRegistry;
 import jakarta.validation.Valid;
