+++ app/src/main/java/com/book/library/dto/DataTransferObject.java
@@ -1,3 +1,3 @@
 package com.book.library.dto;
 
-public sealed interface DataTransferObject permits AnyBookDto, AvailableBookDto, BorrowedBookDto {}
+public sealed interface DataTransferObject permits AnyBookDto, AvailableBookDto, BorrowedBookDto, ReaderDto {}
