+++ app/src/main/java/com/book/library/dto/ReaderDto.java
@@ -1,0 +1,3 @@
+package com.book.library.dto;
+
+public record ReaderDto(int id, String firstName, String lastName) implements DataTransferObject {}
