--- app/src/main/java/com/book/library/reader/ReaderDto.java
@@ -1,3 +1,0 @@
-package com.book.library.reader;
-
-public record ReaderDto(int id, String firstName, String lastName) {}
