+++ app/src/main/java/com/book/library/reader/ReaderRepository.java
@@ -1,5 +1,6 @@
 package com.book.library.reader;
 
+import com.book.library.dto.ReaderDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +10,7 @@ public interface ReaderRepository extends JpaRepository<Reader, Integer> {
 
     @Query(
             """
-                    SELECT new com.book.library.reader.ReaderDto(r.id, r.firstName, r.lastName)
+                    SELECT new com.book.library.dto.ReaderDto(r.id, r.firstName, r.lastName)
                     FROM Reader r""")
     Page<ReaderDto> findAllReaders(Pageable pageable);
 }
