+++ app/src/test/java/com/book/library/book/BookServiceTest.java
@@ -4,8 +4,8 @@ import static org.mockito.ArgumentMatchers.*;
 
 import com.book.library.dto.AvailableBookDto;
 import com.book.library.dto.BorrowedBookDto;
+import com.book.library.dto.ReaderDto;
 import com.book.library.reader.Reader;
-import com.book.library.reader.ReaderDto;
 import com.book.library.reader.ReaderRepository;
 import java.util.Optional;
 import java.util.stream.Stream;
